full system code. needs motor handler functions
uploaded again because I accidentally didn't do them in a folder last time. This code needs the motor handler functions implemented. I also didn't do the HC code for now but just am using the buttons. Also I couldn't figure out how to track whether we were turning more left or right in the start, so instead it is just tracking which sensor is flagged first when driving to the diagonal line after the T-zone.